Fixed bug which makes 'assimilation' appear as 52% on image
As can be seen here https://9axes.github.io/results.html?i=95 and here https://9axes.github.io/results.html?i=5, the 'assimilation' value on the image exported always appears as 52%. This is because the assimilation variable is stored in the global variable 'i', but, in line 180, the for loop overwrites the value of 'i', making it 52 instead. I have fixed this by renaming the 'i' for 'assimilation' to 'i1' (along with all the uses of that 'i'). After testing the changes locally, this bug has been fixed.